retail/requirements.txt
gradio>=4.0.0
pandas>=2.0.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.18.0
pyarrow>=14.0.0
